Add example transformation test case.
You should note that the transformations introduced in this commit
will be refactored heavily in the near future and they don't represent
the final transformation forms yet. At this point we simply need a
test case to see that we don't break anything when we start converging
the transformations towards the final form.

Signed-off-by: Rami Ylimäki <rami.ylimaki@vincit.fi>