Macro _WCL_ADDR_CHECK_* now does only checking, _WCL_ADDR_CLAMP_* does clamping
The new function WebCLConfiguration::getNameOfLimitClampMacro now
returns the name of the macro that does the clamping, previously it
was called by the name returned by
WebCLConfiguration::getNameOfLimitCheckMacro. The latter macro now
returns the name of the function that only checks the limit and
returns a boolean. Both of the macros are generated by
WebCLTransformer::getWclAddrCheckMacroDefinition.