WebCLImageSafetyHandler: added a pass that checks that image2d_t is safely used
When a function with image2d_t argument is called it is checked that
the call arguments always come directly from the parameters of the
current function (this handles passing user-generated data). Also
values of type image2d_t cannot be used if they are not an argument to
a function (this takes care of mutation of the original image
handles).