Add check for using unsupported types as kernel args to WebCLRestrictor
WebCLHeader errored on not being able to figure out a host type for
these, which was equivalent but more obscure.

That was the last WebCLHeader error-out, now it only handles output
after successful validation and can drop the WebCLReporter dependency.

Next issue to tackle is the WebCLTypes dependency and how it implies a
clang::CompilerInstance dependency, which we definitely don't want
outside the library.