WebCLFunctionCallHandler: generalized from WebCLBuiltinHandler
Now the wrapping operation is performed for all function calls, not
just builtin calls , if applicaple.

The matching of operations to perform is done by function call
expression instead of only its name part. (That is, also argument type
can be considered.) This reduces the overall efficiency a bit as the
list of wrapper functions needs to be iterated instead of accessed
directly by the name, but it also adds the ability to perform function
call wrapping by some criteria other tha the function name. Such would
be the case of wrapping all function calls involving sampler_t.