Produce a header from identifiers found in preprocessing, include it later
These identifiers include a lot of stuff, but most usefully names of all
functions that are called by the source. We can leverage this to
generate forward declarations for builtin functions that might be used
by the program, so it's parsed correctly.

Currently the header just includes #warning directives about stuff we
might need to declare, however.