Add back math etc builtin function declarations, but only if used
This is done by matching identifier tokens found by the lexer when
preprocessing against known builtin function names, and emitting
the correct _CL_DECLARE_... macro invocations for those which
match.

Note that the original _CL_DECLARE_... macros from POCL are used, but
only a few of them are usually invoked, with few parameters.

This reverts commit 5be1033f745f944b9484996b30a3394c16ecbe65.