Add back half versions of vload/vstore vector function decls, as needed
This is tested by adding CHECK-NOT to vload-vstore-variations.cl test
for implicit declaration warnings. The half vload/vstore function
invocations in this test are fixed to comply with OpenCL so they don't
error when the prototypes are declared. Previously they were incorrect
but without declarations there was no way to tell.

This reverts commit f1f186d301f6ea358136eaf7c8de01e4c8ad8ff1.

Conflicts:
	lib/kernel.cl