Incredible optimization: Declare only two fake vers of a_w_g(_s)?_c each
This makes the test suite run in 45s instead of 70s on my setup.

It doesn't matter if we have the correct overloads or just ones taking
void * (which everything casts to) as all overloads are forbidden by
name in WebCLAnalyser. We can get there with just the void * dummies.