Implemented OpenCL image access qualifiers using the Clang3.4 interface.
As Clang system doesn't allow retrieving robustly the image2d_t
qualifiers from a plain variable declaration (it doesn't see the
qualifiers through typedefs) a function WebCLTypes::imageKind was
added to do that (it traverses through typedefs). The types
PointerKind and ImageKind were also moved to WebCLTypes to accompany
the new function.

Also a check was added that code like:

  typedef __read_only image2d_t ro_image;
  typedef __write_only ro_image wr_image;

  __kernel void k(wr_image image)

will not be passed by preventing using qualifiers for typedefs that
resolve to image2d_t as kernel arguments. The typedef itself is not
blocked, but blocking that would remove the need to do the check at
kernel argument construction site. Also other uses of that typedef are
not prevented.

The test case was revised to not use qualifiers on typedefs, but a new
test case was added to check that the error is triggered in that case.