Create the clang::DiagnosticConsumer for error reporting by ourselves
It's still one that just prints to the console, but now we can replace
it with our own implementation

The diagnostic options end up slightly different from those in the
internal clang instance, but they're close enough. However, one test
required relaxing a warning check a bit because of this.