WebCLProcessor: added a way to retrieve the list of used extensions
The list of used extensions can only be a subset of the allowed
extensions.

The mechanism for retrieving the used extensions is that a pointer to
the return value is passed down to the WebCLPreprocessor, which then
fills in the value (if the pointer is set). This allows to pierce
through the layers (where the actual instance of WebCLPreprocessor is
not available for us) with the least fuss.

In future this could be refined to support more parameters by
replacing both "extensions" and "usedExtensions" with a more general
environment object.